server/requirements.txt
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Faker==14.2.0
flask-restful==0.3.9
sqlalchemy-serializer==1.4.1
